move persistent queue into its own package
