--- server/queue/README.md
@@ -1,12 +1,0 @@
-# queue package
-
-Go package provides a common interface for working with task queues.
-
-## History
-
-This was originally published in: https://github.com/cncd/queue
-Then it was included in: https://github.com/drone-ci/drone/cncd/queue
-
-## Documentation:
-
-https://godoc.org/github.com/woodpecker-ci/woodpecker/server/queue
