+++ server/queue/persistent.go
@@ -1,37 +1,20 @@
-// Copyright 2021 Woodpecker Authors
-// Copyright 2018 Drone.IO Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package model
+package queue
 
 import (
 	"context"
 
 	"github.com/rs/zerolog/log"
 
-	"github.com/woodpecker-ci/woodpecker/server/queue"
+	"github.com/woodpecker-ci/woodpecker/server/model"
 )
-
-// TODO: move code to "github.com/woodpecker-ci/woodpecker/server/queue"
 
 // WithTaskStore returns a queue that is backed by the TaskStore. This
 // ensures the task Queue can be restored when the system starts.
-func WithTaskStore(q queue.Queue, s TaskStore) queue.Queue {
+func WithTaskStore(q Queue, s model.TaskStore) Queue {
 	tasks, _ := s.TaskList()
-	var toEnqueue []*queue.Task
+	var toEnqueue []*Task
 	for _, task := range tasks {
-		toEnqueue = append(toEnqueue, &queue.Task{
+		toEnqueue = append(toEnqueue, &Task{
 			ID:           task.ID,
 			Data:         task.Data,
 			Labels:       task.Labels,
@@ -45,13 +28,13 @@ func WithTaskStore(q queue.Queue, s TaskStore) queue.Queue {
 }
 
 type persistentQueue struct {
-	queue.Queue
-	store TaskStore
+	Queue
+	store model.TaskStore
 }
 
 // Push pushes a task to the tail of this queue.
-func (q *persistentQueue) Push(c context.Context, task *queue.Task) error {
-	q.store.TaskInsert(&Task{
+func (q *persistentQueue) Push(c context.Context, task *Task) error {
+	q.store.TaskInsert(&model.Task{
 		ID:           task.ID,
 		Data:         task.Data,
 		Labels:       task.Labels,
@@ -65,10 +48,10 @@ func (q *persistentQueue) Push(c context.Context, task *queue.Task) error {
 	return err
 }
 
-// Push pushes multiple tasks to the tail of this queue.
-func (q *persistentQueue) PushAtOnce(c context.Context, tasks []*queue.Task) error {
+// PushAtOnce pushes multiple tasks to the tail of this queue.
+func (q *persistentQueue) PushAtOnce(c context.Context, tasks []*Task) error {
 	for _, task := range tasks {
-		q.store.TaskInsert(&Task{
+		q.store.TaskInsert(&model.Task{
 			ID:           task.ID,
 			Data:         task.Data,
 			Labels:       task.Labels,
@@ -86,7 +69,7 @@ func (q *persistentQueue) PushAtOnce(c context.Context, tasks []*queue.Task) err
 }
 
 // Poll retrieves and removes a task head of this queue.
-func (q *persistentQueue) Poll(c context.Context, f queue.Filter) (*queue.Task, error) {
+func (q *persistentQueue) Poll(c context.Context, f Filter) (*Task, error) {
 	task, err := q.Queue.Poll(c, f)
 	if task != nil {
 		log.Debug().Msgf("pull queue item: %s: remove from backup", task.ID)
@@ -108,7 +91,7 @@ func (q *persistentQueue) Evict(c context.Context, id string) error {
 	return err
 }
 
-// Evict removes a pending task from the queue.
+// EvictAtOnce removes a pending task from the queue.
 func (q *persistentQueue) EvictAtOnce(c context.Context, ids []string) error {
 	err := q.Queue.EvictAtOnce(c, ids)
 	if err == nil {
